+++ src/root_finding/herons_method.rs
@@ -1,4 +1,4 @@
-pub fn herons_method(a: f64, x_0: f64, n_max: i64) -> f64 {
+pub fn herons_method(a: f64, x_0: f64, n_max: usize) -> f64 {
     let mut x: f64 = x_0;
     for _ in 0..n_max {
         x = 0.5 * (x + a/x);
